Reverting float to int for timeout, timeout implementation requires seconds to be an integer value

+++ python/inline/inline.py
@@ -6,7 +6,7 @@ class Here:
         repeated: int = 1,
         tag=[],
         disabled = False,
-        timeout: float = -1.0
+        timeout: int = -1
         # TODO: add a parameter "timeout"
     ):
         """
